Extract Qdrant liveness adapters
Document daemon startup entry points, parameterize permanent-error
coverage, and isolate filesystem and gRPC mechanics behind the private
Qdrant adapter boundary.
